+++ src/main/java/com/masbytes/rbacapi/shared/infrastructure/security/AppUserDetails.java
@@ -1,0 +1,61 @@
+package com.masbytes.rbacapi.shared.infrastructure.security;
+
+import com.masbytes.rbacapi.appuser.domain.entity.AppUser;
+import com.masbytes.rbacapi.shared.domain.enums.Status;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.stream.Collectors;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+
+public class AppUserDetails implements UserDetails {
+
+    private final AppUser user;
+
+    public AppUserDetails(AppUser user) {
+        this.user = user;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        if (user.getRoles() == null || user.getRoles().isEmpty()) {
+            return Collections.emptyList();
+        }
+
+        return user.getRoles().stream()
+                .map(role -> new SimpleGrantedAuthority(role.getRoleName()))
+                .collect(Collectors.toSet());
+    }
+
+    @Override
+    public String getPassword() {
+        return user.getPasswordHash();
+    }
+
+    @Override
+    public String getUsername() {
+        return user.getEmail();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return user.getAppUserStatus() == Status.ACTIVE;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return user.getAppUserStatus() == Status.ACTIVE;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return user.getAppUserStatus() == Status.ACTIVE;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return user.getAppUserStatus() == Status.ACTIVE;
+    }
+}